More controlled
Made the edit such that if the index is equal to -1 the segment will be ligated at the end. Otherwise, it will be ligated at a specific position according to the algorithm in the function as long as the index is not negative and not longer than the sequence itself.